Fix Size Typing
Stores sizes as uint64_t to avoid possible truncation.

llvm-svn: 138901

+++ llvm/lib/CodeGen/AsmPrinter/AsmPrinter.cpp
@@ -1575,7 +1575,7 @@ static void EmitGlobalConstantArray(const ConstantArray *CA, unsigned AddrSpace,
     int Value = isRepeatedByteSequence(CA, AP.TM);
 
     if (Value != -1) {
-      unsigned Bytes = AP.TM.getTargetData()->getTypeAllocSize(CA->getType());
+      uint64_t Bytes = AP.TM.getTargetData()->getTypeAllocSize(CA->getType());
       AP.OutStreamer.EmitFill(Bytes, Value, AddrSpace);
     }
     else {
